поправил архитектуру данных. Вложенные атриббуты сделал для character_template. стилизовал загрузку картинок, сменил стиль полей(не до конца).

+++ db/migrate/20160703112656_create_armors.rb
@@ -2,7 +2,7 @@ class CreateArmors < ActiveRecord::Migration
   def change
     create_table :armors do |t|
       t.string :type
-      t.integer :percent
+      t.float :percent
       t.integer :character_id
 
       t.timestamps null: false
