+++ db/migrate/20160703112750_create_weeknesses.rb
@@ -2,7 +2,7 @@ class CreateWeeknesses < ActiveRecord::Migration
   def change
     create_table :weeknesses do |t|
       t.string :type
-      t.integer :percent
+      t.float :percent
       t.integer :character_id
 
       t.timestamps null: false
