+++ db/migrate/20160703113355_create_armor_templates.rb
@@ -2,7 +2,7 @@ class CreateArmorTemplates < ActiveRecord::Migration
   def change
     create_table :armor_templates do |t|
       t.string :type
-      t.integer :percent
+      t.float :percent
       t.integer :character_template_id
 
       t.timestamps null: false
